Experimenting around directory permissions
The POSIX directory permissions are a little tricky at first: in this asciicast I show how to use a prepared Docker image with two configured users to give a try to them.

The x bit in a directory indicates the traversal permission: if you don't have it you can't access the content of the directory (i.e. the files).

The r bit is instead the read permission: it allows you to read the directory content, the files contained in it.

The x bit has priority: if you can read the name of the files but cannot traverse the directory, you cannot access the files themself.

The Dockerfile ( https://docs.docker.com/ ) used in this video is here ( https://gist.github.com/gipi/d4b9dbf1ad448b98c973 ) .

The command issued are

$ sudo docker build --tag=gipi/chmod https://gist.githubusercontent.com/gipi/d4b9dbf1ad448b98c973/raw/72209d984826680ed788e3e84e285ee82e554206/Dockerfile $ sudo docker run -i -t gipi/chmod /bin/bash